requirements.txt
PyGithub
PyYaml
web-browser